+++ src/movelist.h
@@ -6,7 +6,7 @@
 typedef struct MoveListNode {
     short x;
     short y;
-    int score;
+    float score;
     struct MoveListNode* next;
 } MoveListNode;
 
@@ -18,7 +18,7 @@ typedef struct MoveList {
 
 void List_Init(MoveList** list);
 MoveListNode* List_GetMove(MoveList* list, short index);
-void List_Append(MoveList* list, short x, short y, int score);
+void List_Append(MoveList* list, short x, short y, float score);
 void List_Clear(MoveList* list);
 int List_FindMax(MoveList* list);
 #endif //LIST_H
